added a auth/protectedroutes

+++ backend/controllers/goalController.js
@@ -2,6 +2,7 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //desc Get goals
 //route  GET /api/goals
@@ -41,6 +42,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not Found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user) {
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true,})
 
     res.status(200).json(updatedGoal)
@@ -55,6 +70,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user) {
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.deleteOne()
